Split Issuer in separate file, simplified encryption spec validation, unit testing
