+++ src/main/kotlin/eu/europa/ec/eudi/openid4vci/Types.kt
@@ -19,11 +19,8 @@ import com.nimbusds.jose.EncryptionMethod
 import com.nimbusds.jose.JWEAlgorithm
 import com.nimbusds.jose.JWSAlgorithm
 import com.nimbusds.jose.jwk.JWK
-import com.nimbusds.jwt.JWT
 import com.nimbusds.oauth2.sdk.`as`.ReadOnlyAuthorizationServerMetadata
-import eu.europa.ec.eudi.openid4vci.internal.ProofSerializer
 import io.ktor.client.*
-import kotlinx.serialization.Serializable
 import java.net.URI
 import java.security.cert.X509Certificate
 
@@ -112,40 +109,12 @@ data class CNonce(
  * An identifier of a Deferred Issuance transaction.
  *
  * @param value The identifier's value
- * @param interval The minimum amount of time in seconds to wait before using this identifier to the Deferred Credential Endpoint.
  */
-data class TransactionId(
-    val value: String,
-
-) {
+@JvmInline
+value class TransactionId(val value: String) {
     init {
         require(value.isNotEmpty()) { "Value cannot be empty" }
     }
-}
-
-/**
- * Sealed hierarchy of the proofs of possession that can be included in a credential issuance request. Proofs are used
- * to bind the issued credential to the credential requester. They contain proof of possession of a bind key that can be
- * used to cryptographically verify that the presenter of the credential is also the holder of the credential.
- */
-@Serializable(ProofSerializer::class)
-sealed interface Proof {
-
-    /**
-     * Proof of possession is structured as signed JWT
-     *
-     * @param jwt The proof JWT
-     */
-    @JvmInline
-    value class Jwt(val jwt: JWT) : Proof
-
-    /**
-     * Proof of possession is structured as a CWT
-     *
-     * @param cwt The proof CWT
-     */
-    @JvmInline
-    value class Cwt(val cwt: String) : Proof
 }
 
 /**
