+++ src/main/kotlin/eu/europa/ec/eudi/openid4vci/internal/Serialization.kt
@@ -16,8 +16,12 @@
 package eu.europa.ec.eudi.openid4vci.internal
 
 import com.nimbusds.jwt.SignedJWT
-import eu.europa.ec.eudi.openid4vci.*
+import eu.europa.ec.eudi.openid4vci.Claim
+import eu.europa.ec.eudi.openid4vci.ClaimName
+import eu.europa.ec.eudi.openid4vci.Namespace
+import eu.europa.ec.eudi.openid4vci.ProofType
 import eu.europa.ec.eudi.openid4vci.internal.formats.MsoMdoc
+import eu.europa.ec.eudi.openid4vci.internal.issuance.Proof
 import kotlinx.serialization.*
 import kotlinx.serialization.descriptors.PrimitiveKind
 import kotlinx.serialization.descriptors.PrimitiveSerialDescriptor
