+++ src/main/kotlin/eu/europa/ec/eudi/openid4vci/internal/formats/Format.kt
@@ -21,6 +21,7 @@ import com.nimbusds.jose.jwk.JWK
 import eu.europa.ec.eudi.openid4vci.*
 import eu.europa.ec.eudi.openid4vci.internal.credentialoffer.CredentialSupportedDisplayTO
 import eu.europa.ec.eudi.openid4vci.internal.credentialoffer.LogoObject
+import eu.europa.ec.eudi.openid4vci.internal.issuance.Proof
 import eu.europa.ec.eudi.openid4vci.internal.issuance.RequestedCredentialResponseEncryption
 import kotlinx.serialization.ExperimentalSerializationApi
 import kotlinx.serialization.SerialName
