+++ src/main/kotlin/eu/europa/ec/eudi/openid4vci/internal/formats/W3CSignedJwt.kt
@@ -18,6 +18,7 @@ package eu.europa.ec.eudi.openid4vci.internal.formats
 import eu.europa.ec.eudi.openid4vci.*
 import eu.europa.ec.eudi.openid4vci.internal.credentialoffer.ClaimTO
 import eu.europa.ec.eudi.openid4vci.internal.credentialoffer.CredentialSupportedDisplayTO
+import eu.europa.ec.eudi.openid4vci.internal.issuance.Proof
 import eu.europa.ec.eudi.openid4vci.internal.issuance.RequestedCredentialResponseEncryption
 import kotlinx.serialization.Required
 import kotlinx.serialization.SerialName
